requirements.txt
langchain[openai]
langchain-openai
langchain-chroma
langchain-tavily
python-dotenv
langgraph
langgraph-supervisor
langchain-community
arxiv
wikipedia
flask
beautifulsoup4
markdown
pypdf